requirement.txt
torch==2.0.0
torchvision==0.15.1
Pillow==9.3.0
geopy==2.3.0
opencv-python==4.5.5.64
Flask==2.2.2
googlemaps==4.10.0